+++ src/Exception/Dependency/InvalidBuiltinTypeException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phetit\DependencyInjection\Exception\Dependency;
+
+/**
+ * This exception is thrown when a class's dependency fails to be resolved because builtin type
+ */
+class InvalidBuiltinTypeException extends InvalidTypeException
+{
+}