+++ tests/Fixtures/ServiceWithDependency.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phetit\DependencyInjection\Tests\Fixtures;
+
+class ServiceWithDependency
+{
+    public function __construct(public Service $service)
+    {
+    }
+}