+++ tests/Fixtures/ServiceWithNullableDependency.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phetit\DependencyInjection\Tests\Fixtures;
+
+class ServiceWithNullableDependency
+{
+    public function __construct(public ?Service $service)
+    {
+    }
+}