+++ tests/Fixtures/ServiceWithoutConstructor.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace Phetit\DependencyInjection\Tests\Fixtures;
 
-class Service
+class ServiceWithoutConstructor extends Service
 {
 }